refactor(user): use proper entry point
