+++ apps/io-func-sign-user/CreateFilledDocument/function.json
@@ -14,5 +14,6 @@
       "name": "$return"
     }
   ],
-  "scriptFile": "../dist/create-filled-document.js"
+  "scriptFile": "../dist/main.js",
+  "entryPoint": "CreateFilledDocument"
 }
