+++ apps/io-func-sign-user/package.json
@@ -18,6 +18,7 @@
     "@azure/storage-queue": "^12.11.0",
     "@internal/io-sign": "*",
     "@internal/pdv-tokenizer": "*",
+    "@internal/pdf-handler": "*",
     "@pagopa/handler-kit": "^0.4.2",
     "@pagopa/ts-commons": "^10.10.0",
     "date-fns": "^2.29.3",
