+++ apps/io-func-sign-user/src/infra/azure/functions/fill-document.ts
@@ -2,18 +2,12 @@ import { createHandler } from "@pagopa/handler-kit";
 import * as azure from "@pagopa/handler-kit/lib/azure";
 
 import * as TE from "fp-ts/lib/TaskEither";
-import * as E from "fp-ts/lib/Either";
 
 import { pipe, flow, identity, constVoid } from "fp-ts/lib/function";
 
-import {
-  createPdvTokenizerClient,
-  PdvTokenizerClientWithApiKey,
-} from "@internal/pdv-tokenizer/client";
+import { PdvTokenizerClientWithApiKey } from "@internal/pdv-tokenizer/client";
 import { makeGetFiscalCodeBySignerId } from "@internal/pdv-tokenizer/signer";
 import { ContainerClient } from "@azure/storage-blob";
-
-import { getConfigFromEnvironment } from "../../../app/config";
 
 import { makeFetchWithTimeout } from "../../http/fetch-timeout";
 
@@ -45,34 +39,15 @@ const makeFillDocumentHandler = (
   return createHandler(decodeQueueMessage, fillDocument, identity, constVoid);
 };
 
-const configOrError = pipe(
-  getConfigFromEnvironment(process.env),
-  E.getOrElseW(identity)
-);
-
-if (configOrError instanceof Error) {
-  throw configOrError;
-}
-
-const config = configOrError;
-
-const pdvTokenizerClient = createPdvTokenizerClient(
-  config.pagopa.tokenizer.basePath,
-  config.pagopa.tokenizer.apiKey
-);
-
-const filledContainerClient = new ContainerClient(
-  config.azure.storage.connectionString,
-  config.filledModulesStorageContainerName
-);
-
-const fetchWithTimeout = makeFetchWithTimeout();
-
-export const run = pipe(
-  makeFillDocumentHandler(
-    pdvTokenizerClient,
-    filledContainerClient,
-    fetchWithTimeout
-  ),
-  azure.unsafeRun
-);
+export const makeFillDocumentFunction = (
+  pdvTokenizerClient: PdvTokenizerClientWithApiKey,
+  filledContainerClient: ContainerClient
+) =>
+  pipe(
+    makeFillDocumentHandler(
+      pdvTokenizerClient,
+      filledContainerClient,
+      makeFetchWithTimeout()
+    ),
+    azure.unsafeRun
+  );
